+++ local.config.js
@@ -1,5 +1,6 @@
 const config = {
   FCM_KEY: 'BDAxbWE--q8hHkKSus9Bd47ijSG07V_RhL4HmzbrQN0CpEP4L_Hx6SJsfQ75SHmesXWZq7_NIt0X4vRrkR75_yw',
+  ENC_KEY: 'BasEw47ChKo56F8j06C5dg2w#4/257dDr4r6yR',
   firebaseConfig: {
     apiKey: "AIzaSyB_y27g51gE8HwwypZsRExzK8f0HsvEZ6U",
     authDomain: "bushido-2024.firebaseapp.com",